Remove old docs theme
