--- docs/_theme2/section.hbs
@@ -1,145 +1,0 @@
-<section class='pad2 clearfix prose space-bottom4 fill-white contain'>
-
-  <% if (typeof nested === 'undefined') { %>
-  <div class='clearfix'>
-    <h2 class='fl' id='<%- section.namespace %>'>
-      <%- section.name %>
-    </h2>
-    <% if (section.context && section.context.github) { %>
-      <a class='fr fill-darken0 round round pad1x quiet small' href='<%= section.context.github %>'>
-      <span><%= section.context.path %></span>
-      </a>
-    <% } %>
-  </div>
-  <% } %>
-
-  <%= md(section.description) %>
-
-
-  <% if (section.kind === 'class') { %>
-    <div class='code pad1 small round fill-light'><%= signature(section, true) %></div>
-  <% } %>
-
-  <% if (section.augments) { %>
-    <p>
-      Extends
-      <% if (section.augments) { %>
-        <%= autolink(section.name) %>
-      <% } %>
-    </p>
-  <% } %>
-
-  <% if (section.version) { %><div>Version: <%- section.version %></div><% }%>
-  <% if (section.license) { %><div>License: <%- section.license %></div><% }%>
-  <% if (section.author) { %><div>Author: <%- section.author %></div><% }%>
-  <% if (section.copyright) { %><div>Copyright: <%- section.copyright %></div><% }%>
-  <% if (section.since) { %><div>Since: <%- section.since %></div><% }%>
-
-  <% if (section.params) { %>
-    <div class='pad1y quiet space-top2 prose-big'>Parameters</div>
-    <div>
-      <% section.params.forEach(function(param) { %>
-        <div class='space-bottom0'>
-          <div>
-            <span class='code strong'><%- param.name%></span> <code class='quiet'>(<%= formatType(param) %><% if (param.default) { %>
-            (default <code><%- param.default %></code>)
-            <% } %>)</code> <%= md(param.description, true) %>
-          </div>
-          <% if (param.properties) { %>
-          <table class='space-top1 space-bottom2 fixed-table small'>
-            <colgroup>
-              <col width='30%' />
-              <col width='70%' />
-            </colgroup>
-            <thead>
-              <tr class='strong fill-light'>
-                <td>Name</td>
-                <td>Description</td>
-              </tr>
-            </thead>
-            <tbody class='space-top1'>
-              <% param.properties.forEach(function(property) { %>
-                <tr>
-                  <td class='break-word'><span class='code strong'><%- property.name %></span> <code class='quiet'><%= formatType(property) %></code>
-                  <% if (property.default) { %>
-                    (default <code><%- property.default %></code>)
-                  <% } %></td>
-                  <td class='break-word'><span><%= md(property.description, true) %></span></td>
-                </tr>
-              <% }) %>
-            </tbody>
-          </table>
-          <% } %>
-        </div>
-      <% }) %>
-    </div>
-  <% } %>
-
-  <% if (section.properties) { %>
-    <div class='pad1y quiet space-top2 prose-big'>Properties</div>
-    <div>
-      <% section.properties.forEach(function(property) { %>
-        <div class='space-bottom0'>
-          <span class='code strong'><%- property.name%></span> <code class='quiet'>(<%= formatType(property) %>)</code>
-          <% if (property.default) { %>
-            (default <code><%- property.default %></code>)
-          <% } %><% if (property.description) {
-	    %>: <%= md(property.description, true) %><%
-	  } %>
-          <% if (property.properties) { %>
-            <ul>
-              <% property.properties.forEach(function(property) { %>
-                <li><code><%- name%></code> <%= formatType(property) %>
-                  <% if (property.default) { %>
-                    (default <code><%- property.default %></code>)
-                  <% } %>
-                  <%= md(property.description) %></li>
-              <% }) %>
-            </ul>
-          <% } %>
-        </div>
-      <% }) %>
-    </div>
-  <% } %>
-
-  <% if (section.returns) { %>
-    <% section.returns.forEach(function(ret) { %>
-      <div class='pad1y quiet space-top2 prose-big'>Returns</div>
-      <code><%= formatType(ret) %></code><% if (ret.description) { %>:
-        <%= md(ret.description, true) %>
-      <% }%>
-    <% }) %>
-  <% } %>
-
-  <% if (section.throws) { %>
-    <div class='pad1y quiet space-top2 prose-big'>Throws</div>
-    <ul>
-      <% section.throws.forEach(function(throws) { %>
-        <li><%= formatType(throws) %>: <%= md(throws.description, true) %></li>
-      <% }); %>
-    </ul>
-  <% } %>
-
-  <% if (section.examples) { %>
-    <div class='pad1y quiet space-top2 prose-big'>Example</div>
-    <% section.examples.forEach(function(example) { %>
-      <% if (example.caption) { %><p><%= md(example.caption) %></p><% } %>
-      <pre class='p1 overflow-auto round fill-light'><%= highlight(example.description) %></pre>
-    <% }) %>
-  <% } %>
-
-  <% if (section.members.static && section.members.static.length) { %>
-    <div class='pad1y quiet space-top2 prose-big'>Static Members</div>
-    <%= renderSectionList({ members: section.members.static, renderSection: renderSection, noun: 'Static Member' }) %>
-  <% } %>
-
-  <% if (section.members.instance && section.members.instance.length) { %>
-    <div class='pad1y quiet space-top2 prose-big'>Instance Members</div>
-    <%= renderSectionList({ members: section.members.instance, renderSection: renderSection, noun: 'Instance Member' }) %>
-  <% } %>
-
-  <% if (section.members.events && section.members.events.length) { %>
-    <div class='pad1y quiet space-top2 prose-big'>Events</div>
-    <%= renderSectionList({ members: section.members.events, renderSection: renderSection, noun: 'Event' }) %>
-  <% } %>
-</section>
